disparato/requirements.txt
asgiref==3.5.2
awsebcli==3.20.3
botocore==1.23.54
cement==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.3
future==0.16.0
idna==3.3
jmespath==0.10.0
pathspec==0.9.0
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
tzdata==2022.1
urllib3==1.26.8
wcwidth==0.1.9

Django==4.0
psycopg2-binary==2.9.3
python-datamuse==1.3.2
